feat: config option to run l1 senders in dry run mode
